Double exception reporting unknown field argument type
When the argument to a field is of an unknown type, a schema compilation exception is thrown.

However, creating the message for that exception throws a new exceptions, in certain circumstances.

This was due to the type system changes; I've checked for other incorrect uses of the :type keyword,
and the rest look to be legit.