Fix an issue where error paths in collections were missing indices
When a collection is resolved, and a field resolver of one of its
members raises an error, the spec says that this error _should_
have a key `:path` with an index into that collection.

E.g. in this query:

```GraphQL
query {
  hero {
    friends {
      name
    }
  }
}
```

If the second friend is nil, and name is required, an error with
path `[:hero :friends 1 :name]` should be returned. This index
is added to the `selector-context`, but should have been added
in the `execution-context` inside this `selector-context`.

Please note that this change can potentially add extra errors
to the result because this index breaks the uniqueness of the
error map.