trace namespace depends on dev-only dependencies
I attempted to upgrade to Lacinia 1.1, but production builds do not work because the trace namespace depends on a dev-only dependency (io.aviso/pretty through io.aviso/logging), and the no_unused_fragments pulls in the trace namespace. As a quick fix, remove the unused require; going forward, it looks like io.aviso needs to be included as a non-dev dep. Not sure if this is the path that should be taken, so I am refraining from adding it.